pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.3.0"
description = ""
authors = ["AlexeyLearning <alexeylearning@gmail.com>"]
packages = [
  { include = "page_analyzer" },
]
[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.2"
gunicorn = "^20.1.0"
validators = "^0.20.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
wemake-python-styleguide = "^0.17.0"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
page_analyzer = "page_analyzer.scripts.page_analyzer:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
